requirements.txt
django
djangorestframework
redis
celery
loguru
python-dotenv
psycopg2-binary
django-celery-beat
python-dateutil
pubmed_xml
